Tweaks (#10)
* Adding new layout for skill section

* Adding new tab and table layout for editing skills

* Can create and update skill sections using modal

* Change responsive view of skills

* Removed old resource controller for skill section

* Fixed bug when creating new skill

+++ app/Http/Livewire/Skills/Edit.php
@@ -6,12 +6,13 @@ use Livewire\Component;
 use App\Models\SkillSection;
 use App\Models\Skill;
 use Livewire\WithFileUploads;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class Edit extends Component
 {
 
     use WithFileUploads;
-
+    use AuthorizesRequests;
     public $skill;
     public $is_create;
 
@@ -41,6 +42,9 @@ class Edit extends Component
         $this->skill = $skill;
         $this->is_create = !$skill->exists;
 
+        if (!$this->skill->exists){
+            $this->skill->skill_section_id = SkillSection::firstOrNew()->id;
+        }
         // Validate on page load
         $this->validate();
     }
@@ -68,6 +72,12 @@ class Edit extends Component
          */
         $validatedData = $this->validate();
 
+        if($this->is_create){
+            $this->authorize('create', Skill::class);
+        }else{
+            $this->authorize('update', $this->skill);
+        }
+
         // Save the uploaded logo
         if ($this->uploaded_icon){
             $this->skill->replaceIcon($this->uploaded_icon);
@@ -91,6 +101,7 @@ class Edit extends Component
         /**
          * Delete this given skill
          */
+        $this->authorize('delete', $this->skill);
         $this->skill->delete();
         return redirect()->route('skills.index')->with("message","Skill deleted");
     }
@@ -114,6 +125,7 @@ class Edit extends Component
          * Reset the skill icon
          * to default
          */
+        $this->authorize('update', $this->skill);
         $this->skill->removeIcon();
         session()->flash('info', 'Icon reset to default');
     }
